requirements.txt
fastapi
uvicorn[standard]
openai
llama-cpp-python
transformers
